+++ app/features/adoption/edit/RequiredLabel/index.tsx
@@ -1,4 +1,4 @@
-import { REQUIRED } from '~/features/adoption/create/constants';
+import { REQUIRED } from '~/features/adoption/edit/constants';
 
 export default function RequiredLabel() {
   return <span color="status-active">{REQUIRED}</span>;
