+++ app/models/animal/createAnimal/index.server.tsx
@@ -1,12 +1,12 @@
 import type { Animal } from '@prisma/client';
 import { db } from '~/utils/db/index.server';
 
-export type CreatedAnimal = Omit<
+export type EditingAnimal = Omit<
   Animal,
   'code' | 'openAt' | 'createdAt' | 'updatedAt'
 >;
 
-export default async (animal: CreatedAnimal): Promise<Animal> => {
+export default async (animal: EditingAnimal): Promise<Animal> => {
   const record = await db.animal.create({
     data: animal
   });
